Working

+++ Assets/Scripts/Entity.cs
@@ -1,13 +1,14 @@
 ﻿using UnityEngine;
+using UnityEngine.Events;
 
 namespace DefaultNamespace
 {
     public abstract class Entity : MonoBehaviour
     {
         public abstract float Health { get; }
-        public abstract float MaxHealth { get;  }
+        public abstract float MaxHealth { get; }
         public abstract void Die();
-        public abstract void TakeDamage(int damage);
-        public abstract void Heal(int points);
+        public abstract void TakeDamage(float damage);
+        public abstract void Heal(float points);
     }
 }
